+++ CHANGELOG.md
@@ -275,4 +275,4 @@
 * Define `interruptable` method for `Continuation`
 * Define `Catchable` typeclass to catch errors from `Conclusion`
 * Define `repeat` method of `Stream` datatype
-* Define `subset` method of `Set` interface
+* Define `subset` and `cartesian` methods of `Set` interface